filter_radius is now filter_segments_by_radius for naming consistency.
#26 Refactor post-processors to use new data structure.
